fieat: final changes

+++ Game/Scripts/Multiplayer/Command.cs
@@ -6,13 +6,13 @@ public enum CommandType { Move, Shoot }
 [Serializable]
 public class Command
 {
-    public int Sequence;
+    public ulong Sequence;
     public long EntityId;
     public CommandType Type;
     public Vector2? MoveDir;
 
     public Command() { }
-    public Command(int seq, long eid, CommandType type, Vector2? dir = null)
+    public Command(ulong seq, long eid, CommandType type, Vector2? dir = null)
     {
         Sequence = seq;
         EntityId = eid;
